// server/src/types/socket.enum.ts
export enum SocketEventEnum {
  boardsJoin = "boards:join",
  boardsLeave = "boards:leave",
  boardsUpdate = "boards:update",
  boardsUpdateSuccess = "boards:updateSuccess",
  boardsUpdateFailure = "boards:updateFailure",
  boardsDelete = "boards:delete",
  boardsDeleteSuccess = "boards:deleteSuccess",
  boardsDeleteFailure = "boards:deleteFailure",
  columsCreate = "columns:create",
  columsCreateSuccess = "columns:createSuccess",
  columsCreateFailure = "columns:createFailure",
  columsDelete = "columns:delete",
  columsDeleteSuccess = "columns:deleteSuccess",
  columsDeleteFailure = "columns:deleteFailure",
  columsUpdate = "columns:update",
  columsUpdateSuccess = "columns:updateSuccess",
  columsUpdateFailure = "columns:updateFailure",
  taskCreate = "task:create",
  taskCreateSuccess = "task:createSuccess",
  taskCreateFailure = "task:createFailure",
  taskUpdate = "task:update",
  taskUpdateSuccess = "task:updateSuccess",
  taskUpdateFailure = "task:updateFailure",
  taskDelete = "task:delete",
  taskDeleteSuccess = "task:deleteSuccess",
  taskDeleteFailure = "task:deleteFailure",
}